requirements.txt
# IG Reports Bot - Python Dependencies

# Core dependencies
requests>=2.31.0          # HTTP requests for scraping
beautifulsoup4>=4.12.0    # HTML parsing
lxml>=4.9.0               # XML/HTML parser (faster than html.parser)

# OpenAI LLM
openai>=1.0.0             # OpenAI API client

# Bluesky/AT Protocol
atproto>=0.0.50           # Bluesky posting

# RSS Feed Generation
feedgen>=1.0.0            # RSS/Atom feed generation

# Configuration
python-dotenv>=1.0.0      # Load environment variables from .env

# Database (SQLite is built-in to Python)
# No additional package needed

# Testing
pytest>=7.4.0             # Testing framework
pytest-cov>=4.1.0         # Coverage reporting
pytest-mock>=3.11.1       # Mocking utilities

# Development
black>=23.0.0             # Code formatting
flake8>=6.0.0             # Linting
mypy>=1.5.0               # Type checking

# Optional: Retry logic
tenacity>=8.2.0           # Retry decorators

# Optional: Better logging
colorlog>=6.7.0           # Colored terminal output
